feat: remove stats context, changed to use hook instead
