+++ src/navigation/index.tsx
@@ -10,7 +10,6 @@ import { Content } from "./Content";
 import { Providers } from "../context/composeProviders";
 import { DrawerContextProvider } from "../context/drawer";
 import { CampaignConfigsStoreContextProvider } from "../context/campaignConfigsStore";
-import { StatisticsContextProvider } from "../components/DailyStatistics/DailyStatisticsScreen";
 
 const App = (): ReactElement => {
   return (
@@ -24,8 +23,7 @@ const App = (): ReactElement => {
             HelpModalContextProvider,
             AlertModalContextProvider,
             ImportantMessageContextProvider,
-            DrawerContextProvider,
-            StatisticsContextProvider
+            DrawerContextProvider
           ]}
         >
           <Content />
